refactor(sso): extract shared refreshAccessToken helper
hourly.js and structures.js each carried an identical ~15-line
refreshToken implementation: instantiate SingleSignOn from the EVE_*
env vars, exchange the refresh token, decode the JWT, split sub for
characterID, flatten the scope, then UPDATE hangar.token by id. Lift
it into src/tokenRefresh.js so both scripts (and any future cron jobs)
share one definition.
